Changed static num_workgroups to Int64 as KernelAbstractions does not support Int32)

+++ Core/GPUAlgorithms/Interactions/calculate_interactions.jl
@@ -17,7 +17,7 @@ end #function
 end #function
 
 function calculate_interactions!(θ_updates, cells_data, cell_list_params, numerical_params)
-    workgroup_size = Int32(128)
+    workgroup_size = 128
     num_workgroups = cell_list_params.num_cells
     total_num_threads = workgroup_size * num_workgroups
